+++ NEWS.md
@@ -8,6 +8,8 @@
 
 - Added an argument `template` to `gitbook()` and `epub_book()` (thanks, @ThierryO, #570).
 
+- Added an argument `table_css` to `gitbook()` to allow customized css for tables. (thanks, @haozhu233, #642)
+
 ## BUG FIXES
 
 - HTML output formats such as `gitbook` and `html_document2` won't work when only unnumbered parts (i.e., `# (PART\*)`) are used (thanks, @tjmahr, #575).